Remove redundant reload messages
The code in the authoring package that reloads a package index
included logs to say what was happening, but the core API method that
it invokes does the same thing, so they were always generated twice.

For now, stabilize on the idea that the core will log when it needs
to, so that the external code doesn't have to (and the user will know
what is going on if they don't).