Fix a case bug in the help index topic list
Although the code for looking up a topic to go to it was using a case
insensitive topic search, the code that generates the topic index when
it loads the help index file wasn't storing the keys case folded,
which breaks things when the topic in the index file is not all lower
case.

Previously this was only tested by checking that the topic case in
the help file (when following a link) did not matter, and I happened
to just accidentally make my topics all lower case.