Enable external resources to open web pages
This adds an extra section to the help index that contains references
to external help data.

Currently there are two external types of help data available. The
first is a web URL, which will open an external link in your browser,
while the other allows you to open arbitrary files that appear in  one
of your installed Packages.

This works by parsing a list of topics similar to the one used for
internal help files, then later on merging those topics with the main
topic list from the internal help files to come up with one master
topic list.

ALong the way, all of the unique external references are stored in
lists in the index so that any code can easily check at topic lookup
time if that topic is an external resource or not, and if so what
kind.