Add a history data type
Include a new tuple for tracking a history entry in the navigation
history of a help view.

The data holds the package and file of the help file being viewed as
well as the cursor location that the cursor was last at in that file
and the viewport position so that the view can be returned back to
the same position later.