Initial work on adding history navigation
Included is a new internal help method which is meant to wrap all of
the needed functionality for keeping the history of a help view up to
date.

History is tracked by storing in the settings of the help view an
array of HelpData tuples that tell us the exact help file, package,
viewport position and cursor location in the help file at that point
in history.

The idea is that as you follow topics, the current position is saved
before help navigates away, allowing you to come back to this point
from somewhere else and end up where you started.

At the moment the only thing doing anything with history is the method
for displaying a help file, which ensures that whenever it displays a
help file, if it just created a brand new help file it should have a
default empty history entry created to track it's birth.