Enhance topic display to maintain a history
The show_help_topic() method now takes a parameter that tells the
underlying system if it should maintain the topic history for that
help view.

Passing in True causes the method to update the current history entry
to capture the most recent viewport and cursor location before
following the link and adding a new history entry for that position.

Passing in False causes the method to display the topic without
touching the history. This would be used by commands that are
navigating the history directly and thus don't want to mess up the
history just by observing it.

In order to keep everything in sync, the method will assume that you
passed True even if you did not, if it has to create the initial help
view, since in that case updating the history actually causes the
appropriate data structure to be added.