Include new API method to navigate history
The compulsion to name this method flux_capacitor() was very
overwhelming, but there is now a command that is capable of navigating
through the history attached to a view.

This works by first updating the view and cursor location for the
current history entry, then finding the history entry prior or or
after the current one (as appropriate) and displaying that help file.

If that works, the viewport and cursor are restored to the saved
position and the current history position is changed to be what is
currently being viewed.