Enhance the navigation command to follow history
The command for following links and skipping between anchors now has
an extra parameter to tell it that you want to follow history.

This is bound to default keys and triggers the new API method just
created to attempt to move through history.

Although the method will display a warning if you try to navigate in a
direction that is not allowed, currently the command is disabled for a
particular direction if it can't go in that directory, so this is not
seen.