Fix latent bug following a link in authored files
The code for following a link by double clicking works by checking to
see if the user is trying to double click to select text and, if so,
seeing if the item under is a link and trying to follow it.

The command for doing this is smart enough to disable itself if it's
not invoked directly on the help view, but the event listener always
tries to follow the link anyway.

The result of this is that if you double click on a link from inside
of a help file being authored, the selection command will be altered
to be a link follow, which does nothing, which means that not only do
we not follow the link, but also the selection doesn't happen either.

This makes it impossible to double click to select a link while you're
authoring help.

The fix is to only trigger the event when it happens on a view that is
read only (and still matches the scope), since in that case it is more
likely to be the help view.