Make the navigation command a window command
Previously the command for navigating in help was an instance of
TextCommand, which means it gets invoked directly on a view in
response to a command.

The down side to this is that we want to put the new navigation
command for history into the menu, and a part of that is to ensure
that the command disables itself when there is not any history.

If this command remains a text command, then it will always report
itself as being disabled unless the cursor happens to be inside of it,
which is not what we want.

Instead, the command is now a WindowCommand that finds the help view
in its window and works with that, disabling itself if the window does
not have a help view at all.

The command is now also smart enough to disable itself in the case
that there IS a help view, but it has no history to navigate to in
whatever direction the command would be invoked for.