Include the history commands in the menu
The command for navigating now has a description method that will tell
you the file that it's going to go to in whichever navigation
directory you specify (assuming it can go in that direction) and the
main menu includes entries for this command as well.

With this in place, the menu will show you if you can move through
history, and if so what topic you will end up at.