Change the history keys for MacOS
Based on the statistical sample of just my Mac, this alters the keys
that are used to move through history to match what Safari, Chrome and
Firefox use for this (super+[ and super+]).

These technically collide with the keys for indenting on MacOS, but
since they're only applied to read only help views that is hopefully
not an issue.