New API function for getting current help file
Since there was already a common routine for getting the current help
package, there is now also a routine for getting the current help file
as well, for orthogonality.

There should probably also be something for wrapping checking if
they are set and setting them as well, to shield everything from the
custom settings used, but baby steps.