For clarity, multi line logs are prefixed
While working on the authoring tools I noticed that when error dialogs
were logged, it was confusing in the console since the prefix is only
applied to the first line of the log.

This modifies the log function so that it displays the prefix on all
of the lines, which if nothing else gives them all the same relative
start position and makes them easier to read.