Allow index reload on non-packaged file
In order to reload a hyperhelp.json file in the authoring tools, the
filename needs to be contained within the Packages folder so that we
can determine the name of the package that is hosting the help.

It's possible not only for an index file to be stored as a packaged
file (the default for most users of the package) but also for the
file to be stored elsewhere but symlinked to a location that is inside
of the Packages folder.

Originally the command for reloading the package index would call
`realpath` on the filename, which would resolve all symbolic links
in order to get at the final file.

Now it will see if the file name as loaded is in the Packages folder
and if so use it, but if not it will resolve the symlink and try
again.

This allows the package to remain outside of the Packages folder but
still be reloaded if it's opened via its symlink.