Start the first phase of code separation
The code layout of the repository is undergoing some changes in order
to split apart the two packages (hyperhelp and the authoring tools)
into two separate subfolders.

Going forward anyone using the code will have to do a partial git
checkout to get at each package individually, or make a check out and
manually copy the appropriate package over.

This just changes the location of the files, but no work has been done
as of yet to actually make this work, so things are likely somewhat
broken, particularly with regards to the authoring tools.