Support the new directory layout in the project
The new layout of the project causes some layout changes in the packed
project file in order to keep everything visible.

For development reasons, it's important to be able to open package
files with paths that are relative to the Packages folder, which means
opening them from a symlink in the Packages folder.

To support this, the original path entry has patterns to exclude the
two new package folders, and then those folders are manually added
back with their Packages folder path.

The net result of this is that the sidebar (if you use it) now shows
three folders; one for each package and one for the top level files.

This is important because it means that files opened will open
relative to the Packages folder but also only one copy is present, so
find in files and other file related tasks won't be looking at
duplicate entries.