Enhance the reload help command
This enhances the code responsible for reloading the currently
displayed help file so that it's more useful.

Previously the command could only be triggered from within the help
view itself, and would cause that file to reload.

Now the command works not only in that case, but also from within an
edit view of the current help file.

This allows you to edit text, then hit the reload immediately to see
the changes in the help view (if it's visible).

This currently keeps the focus on the help view when it triggers,
which may be undesirable. However since the reload command also resets
the viewport back to the top of the file, this is currently handy for
being able to navigate.

Since the command works from any help file now, the context has been
removed from the key bindings.

This may need to be revisited in the future with a setting so that it
doesn't boost the key binding for people that want to use F5 for
something else while editing.