Include nascent hover popup on topics
A hover popup is added on help topics in the help view that will
display the name of the help topic as configured in the index, along
with showing what will happen.

Links to URL's tell you what URL they will open, links to package
files tell you what file they will open, and links to regular help
topics will tell you what file will be navigated to.

Additionally, if a help topic is not known, the popup will tell you
that it's missing.