Include the beginning of a crude linter
This leaves much to be desired in code quality and things will need to
be heavily cleaned up and refactored, but I wanted to get at least a
minimal version working to make finishing the initial help a bit
easier.

Currently the only lint check that is done is to find help links that
refer to targets that don't appear in any help files, which makes it a
little easier to know where unfinished help is.

The command is only enabled while a help index or help file has the
focus, as long as the file is rooted in the package folder and is
for a package whose help index is loaded.

If it's run on an index, all of the help files are linted to see if
they have any missing anchors, while if it's run on a specific help
file only that file is linted.

Output currently always goes to a panel, but this will change in the
future. Adding regions will also be added.