Disable help reload if this file is not help
This is still not all the way there because although the command
should work from any help file, it really should only work if the
edited file is the current help.

I'm leaving this for now because this command will be enhanced so
that using it when editing a help file that's not current will also
cause that help file to be loaded if it's not aready.