Include a new topic lookup global API call
This abstracts away all of the work that has to happen in order to
look up a topic to get at its data.

You must pass in either a help data structure or the name of a
package, plus the topic you want to look up. The lookup returns either
the topic data or None.

The lookup takes care of any topic string modifications that need to
happen in order to find the topic, such as folding the case or
replacing any embedded spaces with tab characters so that the lookup
will work.