Allow aliases in help topics
This updates the validator to allow a key of "aliases" in help topics
inside of help files, whose value should be an array of strings that
represent other topics that also refer to this topic directly.

A list of aliases is optional in a topic, and is currently not allowed
for externals.