Populate aliases into the HelpData structure
When loading help topics from help files, the index loader will now
check to see if there is a list of aliases for this topic, and if
there is it will validate that the alias doesn't already exist as an
alias or as a topic, and then add it to the list of topic aliases.

Similarly, while topics are being handled they're validated as not
being an existing alias as well.

The alias list is a dictionary that keys an alias name directly on the
topic that it aliases, allowing you to provide multiple names for the
same topic.

This lets you do things like have links that are slight variations on
each other.