Allow aliases for topics in the TOC
When specifying the table of contents in the help index, you can
use a topic entry that contains a caption in order to override the
caption that associates with the topic, or optionally leave the
caption out, in which case the topic assigned to the caption is
used instead.

As a shorthand for the second case, you can include the topic as just
a string, and it will be used to look up the topic directly.

This allows aliases to be used in the TOC in both places, so that
aliases are allowed as topics.