Use the new alias functionality in the core help
There are a few places in the core help where links were broken
because the anchor used a pluralized version of the anchor name but
the link text used a singular version (e.g. anchor of `links`, but
with a link to `link` instead).

Using the power of aliases, this allows for those links to resolve
properly.

Also there were some headings using hidden anchors where only one word
of the header was in the anchor so that it could be linked to. These
places now have a full text anchor with aliases to the short versions.