Include the initial version of the SnAPI help
This includes the current state of an external project called SnAPI
into the repository for devember 2017.

SnAPI is a version of the Sublime Text 3 API in the hyperhelp format
of a previous prototype version of hyperhelp.

It's included here to act as an extra stand in for a more completely
functional help system for some new upcoming code.

As this is for a previous version of hyperhelp, the help files and the
index are not in the proper format for this newer version, but that
will be fixed in the next commit.