Update SnAPI to the latest help format
This is the initial workings of the original SnAPI help updated to
work with the new format of files used by this version of hyperhelp.

Primarily this alters the index to be in the new format, although some
changes to help to account for the new ability to have hidden anchors
and aliases have been done as well.

The help content itself is still lacking a bit behind of the current
version of the API since this was originally worked on a little bit
before Sublime Text 3.0 was officially released.

When devember ends and the files here get merged back into the source
repository, that will be fixed.