Include a key binding to open the SnAPI index
At least for the current time while development is happening, this
includes a new default key binding to open the help index for SnAPI.

Currently all platforms use the same key binding.

As a side effect of this, a redundant parameter was removed from the
similar command in the command palette.