Allow help index to customize the default caption
Currently the default caption for any topic that doesn't specify one
of it's own (and which is not inherited) is one that tells you the
topic and its location, allowing you to figure out where it is.

In some cases (for example SnAPI), there may be large numbers of
topics for which the caption can easily be inferred based on the topic
itself.

This adds a new optional top level index key that allows you to
specify the format of the default caption for cases such as this.