Include support for simple code blocks
The syntax for help files has been extended to provide a simplistic
method of using a triple back tick in order to introduce a code block
that spans multiple lines.

Currently this requires the fences to appear on lines of their own,
but this will be tweaked in the future.