Enhance the syntax for anchors
The syntax element for a regular anchor was still "broken" in that it
requires a link to be a minimum of two characters even though really
only one is needed.

The syntax element for a hidden anchor did not have this problem, but
it DID have an issue with allowing characters that it should not.

Both have now been fixed to match more strictly.