Enhanced key bindings for link and anchor edits
The interior of links and anchors requires a tab in order to be
appropriately represented as whitespace, since actual spaces are not
allowed.

By applying some appropriate punctuation scopes to the characters that
define these constructs, there is now a key binding which inserts a
tab if the cursor is in the interior of a link or anchor, so that you
don't have to do anything special in this case.

This can (and should) be enhanced in the future, since it allows for
trailing space characters. This is mostly a place holder to get the
feature nominally in place to make editing help easier in the interim.