Factor the linter code out into its own file
There's not much time left for tonight, but there are some new linter
commands upcoming, so the associated code is best in its own file.

For now the command has been boosted, but probably in a future change
the new file will be pure lint code and the command will invoke some
methods, and so it will go back into the commands file.