Allow the linter to know what kind of lint to do
Currently the only lint command is to validate links, and that can
work either on the current file or on all files, depending on what
file the command is entered in.

In the future there will be many things to lint and which ones get
selected will depends on the source type. For example, validating that
all help files are present only happens when doing a full package
lint, while checking that links are valid can be done file by file.