Stage one of a big lint refactor
This pulls a lot of the linting code out of the lint command itself
and puts it into helper routines where it belongs.

This is just the first part of this operation, which aims to remove
all of the code not directly related to performing the lint so that
more lint types can be added.