Include a new common function to load a resource
This implements a new common function for loading a resource, which
loads the sublime resource as binary and then decodes it as UTF-8,
returning the result after normalizing the line endings to be unix
because that's what Sublime uses internally.

This is in common because both regular help, indexes and the authoring
tools may want to use it for their own purposes, since it's a single
point of entry for loading a resource, validating that it exists, and
verifying that it's valid UTF-8.