Use the new resource loader for help
The code for loading a help file is now much simplified by using the
recently added common function for loading and decoding a resource.

This is still a wrapper because it also constructs the appropriate
resource spec for the loaded file internally, which makes external
code easier to read.

Additionally this will be extended later to automatically fall back
to a local file like the help index does.