Refactor the load_resource method to open locally
The code from the help index that allows it to load an index file from
a local file or from a help resource has been pushed up into the
load_resource() method, so that that method can load a resource of
either type based on the input specification and will return the
appropriate data back.

This makes the help index loader much simpler since it just needs to
determine what the index resource would have been but can otherwise
just make a single call to load the resource.