load_resource() is now more transparent
The code in load_resource() is now more transparent; it always takes
as input a resource specification, but if the resource can't be loaded
by Sublime, it automatically falls back to trying to load the file
manually from a file in the packages folder.

This allows it to work immediately even for resources that have just
been newly added without the loading code having to do anything
special.