help index loader cleanup
Using the new resource loader with automatic fallback, the code for
loading a help index is now greatly cleaned up; it only takes regular
resource specifications, removing a lot of redundant code.

As a result of this change, the authoring command for adding help to a
package now loads the index from a regular resource specification as
using a file spec will not work.

The authoring command can be cleaned up a bit more, but that can
happen during the regular cleaning pass.