Generate stub files in utf-8 for completeness
For the sake of completeness, the code in the authoring tools for
creating stub help and index files now explicitly generate the files
using utf-8.

This is more for orthogonality than anything else, but at this point
the only code that's still not doing this is the code in the linter
that will go away soon.