Include new base class for linting
This introduces a new class LinterBase that will be the base of all
linting operations in the authoring tools. This provides the core of
what's needed in interact with the upcoming changes to the linting
code.

Also included ia new data structure of type LintResult that tracks the
location of an issue, the resulting text, and the type of issue that
the result contains (info, warning or error).

In operation a subclass of this class will be created for each lint
operation that is to be taken out. One by one each target file is
loaded and passed to the lint() method of all linters so they can
gather any information that they require from it.

Once all inputs have been processed, the result() method is invoked to
obtain a list of LintResult tuples that provide any found issues, and
these are turned into a report and displayed.

Depending on the lint operation being undertaken, different aspects of
the class can be overriden. For example simple linters might be able
to do everything from the constructor, others may be able to determine
issues on a file by file basis, and others may require gathering
information from all files before being able to do any checks.