Include new lint format/display code
In order to take advantage of the new linting code, the way that the
output gets created needs to be pushed out of the lint command itself
and re-worked to support the new infrastructure.

There is now a new function that takes a list of results and formats
a report that displays all issues found on a file by file basis.