Refactor the lint command to use the new code
All of the linting code has been pushed up to a new subclass of the
linting base class and the new display routines are used to render the
output.

As a test of the system overall, there is a second linter now which
finds anchors not currently in the index as well. This is 99%
identical to what the broken link linter is doing and will be factored
into there at a later date, but for the time being this is enough to
ensure that things are working as expected.