Make link regions persistent
Instead of having a plugin handler that re-underlines links in active
help views on load, this makes the regions persist in the session file
instead so that Sublime does the heavy lifting, which looks visibly
better.

To be honest I don't think I ever noticed the persistent flag for
regions prior to now.