Include new linter function for getting linters
A new linter function has been created which takes a lint target and
returns back the list of linter instances to be run across that list
of targets.

This is something that used to be done by the command itself, but by
having it external, the command can move back to the same location as
the other commands and still keep all of the linter specific code in
the same place.