Remove redundant panel clearing
Due to a minor brain fart, I forgot that recreating an existing panel
also clears the contents for you, so there is no need for the code
that is displaying the lint results to do this.

The read-only status of the panel is not altered by this call, so we
still need to flip the state of that flag while modifying it.