Clean up the linter API for command move
Now that the linting code is somewhat stabilized, the helper methods
have had their privacy dropped in preparation for moving the lint
command back to the appropriate location.

This also adds a simplification that allows you to tell the function
for formatting the lint report that it should also display it before
it returns, which saves a step.

Probably this could just be renamed to a single thing or the two could
be rolled together, but I want to wait a bit to see if the separation
is useful for anything else first.