Fix broken error message on lint failure
The code for displaying a failure to find and load a lint file when
performing a lint was broken and not referring to the correct package
information when trying to formulate the error, which caused an
exception.

This has now been fixed so that the error message is displayed
correctly.