Include a shortcut for adding index issues
At the current time, linting in the help index doesn't provide an
actual view to the file since without a nice syntax there is no way
except hand parsing the source JSON to know where errors should be
located.

As such, any errors focused on the index always open at the first
character of the file for now, just so that issues can help you open
the index.