Include a help index linter
This is an example of a linter that is only enabled when the target of
the lint is a help index file, so it doesn't get used if you're
linting in a help source file directly in order to not cruft up the
output.

This linter checks that the list of help files in the index is the
same as the list of help files that are present in the package in  the
document root.

If the two lists are not the same, issues are added to tell you when
help files exist in the document root but not the index as well as
when help files exist in the index but not on disk.

Help files that exist but are not indexed are just a warning since
they may not be help files, but help files that are indexed but don't
exist are errors since the system expects them to be there.