Ensure that the index is always clickable in lint
The original version of the lint code assumed that the help index is
stored in the document root. This is the most common case but is not
strictly necessary.

The constructor of the linter base now determines what the appropriate
path to the index is so that it can use that instead. The path
returned is always relative to the document root, which is the base
directory set for results in the lint panel.