Upgrade the core help for hyperhelp
This is still missing some key bits, such as the API and the commands
that are available to external packages, but it's most of the way
there for describing everything else.

Aside from the above the only other things to do are to include some
information on the structure of the index files to provide more help
for authoring.