Expand authoring key bindings
A few commits back a change was added to have the space bar insert a
literal tab when inside of a link or an anchor. The change in this
commit is twofold:

1. Include this functionality for key bindings as well
2. Make the binding available on Windows and MacOS as well

As defined, the context for being applied within a key binding in a
help file is set to also allow operation on the punctuation that ends
a key binding.

At this moment this is not necessary since the punctuation around key
bindings doesn't have a special meta scope applied, but the plan is to
add that in the future, so the binding is ready for it now.