Include a prototype SnAPI command
This adds a plugin command in SnAPI that opens the help index as the
standard HyperHelp command would, but with some extra capabilities.

If the current file is a python file, the command will extract the
selected text out of the buffer (or the word the cursor is on if there
is no selection) and see if it's a help topic.

When it is, the topic is displayed directly. Otherwise the index is
displayed but the text from the buffer is added as the initial filter
text to hopefully narrow down the possible topics, which makes life a
little easier.

This is currently on a temporary key binding until some work is done
to see if there's something more standard that could be used as a
default instead.